refactor(pds.__init__.py and various): Created a top-level Purpose Enum to be used in the module.
